Usuario testado com swagger. Migracao para PostgressSql com banco de dados local via dBeaver

+++ src/main/java/com/projetoPessoal/service/UserService.java
@@ -1,33 +1,56 @@
 package com.projetoPessoal.service;
 
-
+import com.projetoPessoal.exception.UserNotFoundException;
 import com.projetoPessoal.model.User;
 import com.projetoPessoal.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 
-//Lógica de negócio
 @Service
 public class UserService {
 
-    //Implementação dos métodos
     @Autowired
     private UserRepository userRepository;
 
     // Buscar por ID
     public User findById(Long id) {
         return userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado / User not found"));
+                .orElseThrow(() -> new UserNotFoundException(id));
     }
 
-    // Salvar usuário
+    // Listar todos
+    public List<User> findAll() {
+        return userRepository.findAll();
+    }
+
+    // Criar ou atualizar
     public User saveUser(User user) {
         return userRepository.save(user);
     }
 
-    // Deletar usuário
+    // Atualizar usuário existente
+    public User updateUser(Long id, User userDetails) {
+        User user = findById(id);
+
+        user.setName(userDetails.getName());
+        user.setEmail(userDetails.getEmail());
+        user.setPhone(userDetails.getPhone());
+        //user.setAddress(userDetails.getAddress());
+        user.setIncome(userDetails.getIncome());
+        user.setNumOfDependents(userDetails.getNumOfDependents());
+        user.setStatus(userDetails.getStatus());
+        user.setObservations(userDetails.getObservations());
+
+        return userRepository.save(user);
+    }
+
+    // Deletar
     public void deleteById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
 }
